Initial split of crates
