+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["derive", "."]
+members = ["derive", ".", "codetable"]
 
 [package]
 name = "multihash"
@@ -15,26 +15,17 @@ edition = "2021"
 rust-version = "1.59"
 
 [features]
-default = ["std", "derive", "multihash-impl", "secure-hashes"]
-std = ["unsigned-varint/std", "multihash-derive/std", "alloc"]
-alloc = ["core2/alloc"]
-multihash-impl = ["derive"]
-derive = ["multihash-derive"]
+default = ["std", "derive"]
+std = ["unsigned-varint/std", "multihash-derive?/std", "alloc"]
+alloc = []
+derive = ["dep:multihash-derive"]
 arb = ["quickcheck", "rand", "arbitrary"]
-secure-hashes = ["blake2b", "blake2s", "blake3", "sha2", "sha3"]
-scale-codec = ["parity-scale-codec"]
-serde-codec = ["serde", "serde-big-array"]
-
-blake2b = ["blake2b_simd"]
-blake2s = ["blake2s_simd"]
-identity = []
-sha1 = ["digest", "sha-1"]
-sha2 = ["digest", "sha-2"]
-sha3 = ["digest", "sha-3"]
-strobe = ["strobe-rs"]
-ripemd = ["ripemd-rs"]
+scale-codec = ["dep:parity-scale-codec"]
+serde-codec = ["dep:serde", "dep:serde-big-array"]
 
 [dependencies]
+arbitrary = {version = "1.1.0", optional = true }
+core2 = { version = "0.4.0", default-features = false }
 parity-scale-codec = { version = "3.0.0", default-features = false, features = ["derive"], optional = true }
 quickcheck = { version = "1.0.3", optional = true }
 rand = { version = "0.8.5", optional = true, features = ["small_rng"] }
@@ -42,29 +33,9 @@ serde = { version = "1.0.116", optional = true, default-features = false, featur
 serde-big-array = { version = "0.3.2", optional = true, features = ["const-generics"] }
 multihash-derive = { version = "0.8.0", path = "derive", default-features = false, optional = true }
 unsigned-varint = { version = "0.7.1", default-features = false }
-arbitrary = {version = "1.1.0", optional = true }
-
-blake2b_simd = { version = "1.0.0", default-features = false, optional = true }
-blake2s_simd = { version = "1.0.0", default-features = false, optional = true }
-blake3 = { version = "1.2.0", default-features = false, optional = true }
-digest = { version = "0.10.1", default-features = false, optional = true }
-sha-1 = { version = "0.10.0", default-features = false, optional = true }
-sha-2 = { version = "0.10.0", default-features = false, optional = true, package = "sha2" }
-sha-3 = { version = "0.10.0", default-features = false, optional = true, package = "sha3" }
-strobe-rs = { version = "0.7.0", default-features = false, optional = true }
-ripemd-rs = { package = "ripemd", version = "0.1.1", optional = true}
-
-core2 = { version = "0.4.0", default-features = false }
 
 [dev-dependencies]
-criterion = "0.3.3"
 hex = "0.4.2"
 serde_json = "1.0.58"
 quickcheck = "1.0.3"
-rand = "0.8.5"
 arbitrary = "1.1.0"
-multihash = { path = ".", features = ["sha1", "strobe"] }
-
-[[bench]]
-name = "multihash"
-harness = false
