+++ codetable/benches/multihash.rs
@@ -1,11 +1,11 @@
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
-use rand::Rng;
-
-use multihash::{
-    Blake2b256, Blake2b512, Blake2s128, Blake2s256, Blake3_256, Hasher, Keccak224, Keccak256,
-    Keccak384, Keccak512, Sha1, Sha2_256, Sha2_512, Sha3_224, Sha3_256, Sha3_384, Sha3_512,
-    Strobe256, Strobe512,
+use multihash::Hasher;
+use multihash_codetable::{
+    Blake2b256, Blake2b512, Blake2s128, Blake2s256, Blake3_256, Keccak224, Keccak256, Keccak384,
+    Keccak512, Sha1, Sha2_256, Sha2_512, Sha3_224, Sha3_256, Sha3_384, Sha3_512, Strobe256,
+    Strobe512,
 };
+use rand::Rng;
 
 macro_rules! group_digest {
     ($criterion:ident, $( $id:expr => $hash:ident, $input:expr)* ) => {{
